Move `get_many_unchecked_manual_uniqueness()` test from the file with `QueryState` to the file with `Query`.
